requirements.txt
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
cryptography==37.0.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.50.0
django-appconf==1.0.5
django-ckeditor==6.7.0
django-compressor==2.4.1
django-debug-toolbar==3.4.0
django-extensions==3.1.5
django-js-asset==2.1.0
django-import-export
django-phone-field==1.8.0
django-tinymce==4.1.0
django-user-agents==0.4.0
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
idna==2.8
importlib-metadata==4.12.0
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.15.0
jupyter==1.0.0
jupyter-console==6.4.3
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
libsass==0.20.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
more-itertools==8.0.2
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.22.3
oauthlib==3.1.0
openpyxl
packaging==21.3
pandas
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
pyotp
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==6.0
pyzmq==23.2.1
qrcode
qtconsole==4.6.0
rcssmin==1.0.6
requests>=2.24.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
Send2Trash==1.8.0
six==1.16.0
sorl-thumbnail==12.8.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.5.0
terminado==0.15.0
testpath==0.4.4
text-unidecode==1.3
tinycss2==1.1.1
tornado==6.2
traitlets==5.3.0
ua-parser==0.16.1
urllib3==1.26.12
user-agents==2.2.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlwt==1.3.0
zipp==3.8.1
